new instructions on loading GECKO
Any generic changes that are required to the functions I will try to push towards the GECKO repository, so you do not need to make custom scripts that could become out of date. These instructions will change later, when the changed files will have been accepted in a new version of GECKO, but this way you can already work with the intermediate version.